minor enhancements

+++ src/main/java/com/uep/wap/repository/InventoryItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InventoryItemRepository extends CrudRepository<InventoryItem, Integer> {
+public interface InventoryItemRepository extends CrudRepository<InventoryItem, Long> {
 }
