+++ src/main/java/com/uep/wap/repository/InvoiceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InvoiceRepository extends CrudRepository<Invoice, Integer> {
+public interface InvoiceRepository extends CrudRepository<Invoice, Long> {
 }
