+++ src/main/java/com/uep/wap/repository/StockTransferLineRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockTransferLineRepository extends CrudRepository<StockTransferLine, Integer> {
+public interface StockTransferLineRepository extends CrudRepository<StockTransferLine, Long> {
 }
