+++ src/main/java/com/uep/wap/repository/WarehouseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WarehouseRepository extends CrudRepository<Warehouse, Integer> {
+public interface WarehouseRepository extends CrudRepository<Warehouse, Long> {
 }
